+++ src/adapters/cosmosdb/client.ts
